feat: extract bibliography loading into citum-refs
Consolidates parsing into new lower-level crate.
