+++ crates/citum-engine/Cargo.toml
@@ -15,6 +15,7 @@ crate-type = ["rlib", "cdylib"]
 [dependencies]
 citum-schema = { workspace = true, features = ["legacy-convert"] }
 citum-schema-data = { workspace = true }
+citum-refs = { workspace = true }
 csl-legacy = { workspace = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
