+++ crates/citum-refs/src/formats/mod.rs
@@ -1,0 +1,11 @@
+/*
+SPDX-License-Identifier: MIT OR Apache-2.0
+SPDX-FileCopyrightText: © 2023-2026 Bruce D'Arcus and Citum contributors
+*/
+
+//! Format-specific reference data parsing helpers.
+
+pub mod biblatex;
+pub mod csl_json;
+pub mod native;
+pub mod ris;
